+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 Version 1.0 (dev)
 -----------------
 
+* Added caching support for the complete ``{% render_placeholder %}``, ``{% page_placeholder %}`` and ``{% sharedcontent %}`` tags.`, ``{% page_placeholder %}`` and ``{% sharedcontent %}`` tags.
 * Fix handling ``CheckboxSelectMultiple`` in admin form widgets.
 * Fix missing API parameters for ``ContentItem.objects.create_for_placeholder()`` and ``Placeholder.objects.create_for_parent()``.
 * Fix static default ``SITE_ID`` value for ``SharedContent``, for compatibility with django-multisite_.